send access

+++ src/main/java/com/ynov/crm/service/OrganizationServiceImpl.java
@@ -1,20 +1,26 @@
 package com.ynov.crm.service;
 
+import java.util.HashSet;
 import java.util.List;
+import java.util.Vector;
 import java.util.stream.Collectors;
 
 import javax.transaction.Transactional;
 
+import org.apache.tomcat.jni.User;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
+import org.webjars.NotFoundException;
 
+import com.ynov.crm.enties.AppUser;
 import com.ynov.crm.enties.Organization;
 import com.ynov.crm.mapper.OrganisationMapper;
 import com.ynov.crm.repository.AppUserRepository;
 import com.ynov.crm.repository.OrganizationRepository;
 import com.ynov.crm.requestdto.OrganizationRequestDto;
 import com.ynov.crm.responsedto.OrganizationResponsDto;
+import com.ynov.crm.utils.CheckAccessAdmin;
 
 import lombok.Data;
 import lombok.extern.slf4j.Slf4j;
@@ -27,14 +33,23 @@ public class OrganizationServiceImpl implements OrganizationService{
 	private OrganizationRepository organizationRepository;
 	private AppUserRepository userRepository;
 	private OrganisationMapper organisationMapper;
+    private CheckAccessAdmin checkAccessAdmin;
+    private UserPrinciple currentUser;
 	
-	@Autowired
+	
 	public OrganizationServiceImpl(OrganizationRepository organizationRepository, AppUserRepository userRepository,
-			OrganisationMapper organisationMapper) {
+			OrganisationMapper organisationMapper, CheckAccessAdmin checkAccessAdmin) {
 		super();
 		this.organizationRepository = organizationRepository;
 		this.userRepository = userRepository;
 		this.organisationMapper = organisationMapper;
+		this.checkAccessAdmin = checkAccessAdmin;
+		
+
+		
+	}
+	public void initCurrentUser(){
+		 currentUser  =  (UserPrinciple) SecurityContextHolder.getContext().getAuthentication(). getPrincipal();
 	}
 	
 	@Override
@@ -54,15 +69,31 @@ public class OrganizationServiceImpl implements OrganizationService{
 
 	@Override
 	public OrganizationResponsDto getOrganization(String orgId) {
-		return (organizationRepository.existsById(orgId))?
-				organisationMapper.OrganisationToOrganizationResponseDto(organizationRepository.findById(orgId).get()):
-					null;
+		this.initCurrentUser();
+		if(organizationRepository.existsById(orgId)) {
+			Organization  organization= organizationRepository.findById(orgId).get();
+			if(checkAccessAdmin.checkAccess(organization.getAdminId(), currentUser)) {
+				return organisationMapper.OrganisationToOrganizationResponseDto(organization); 
+			}
+		}
+		
+		return new OrganizationResponsDto();
+		
 	}
 
 	@Override
 	public List<OrganizationResponsDto> findAll() {
+		this.initCurrentUser();
 		log.info(organizationRepository.findAll().toString());
-		return (organizationRepository.findAll())
+		List<Organization> organizations = organizationRepository.findAll();
+		List<Organization> newOrganizations = new Vector<>();
+		organizations.forEach(organization->{
+			if(checkAccessAdmin.checkAccess(organization.getAdminId(), currentUser)) {
+				newOrganizations.add(organization);
+			}
+	
+		});
+		return newOrganizations
 				.stream()
 				.map(organization -> organisationMapper.OrganisationToOrganizationResponseDto(organization))
 				.collect(Collectors.toList())
@@ -71,9 +102,12 @@ public class OrganizationServiceImpl implements OrganizationService{
 
 	@Override
 	public OrganizationResponsDto save(OrganizationRequestDto organizationRequestDto) {
-		UserPrinciple user =  (UserPrinciple) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+		this.initCurrentUser();
+		AppUser user = userRepository.findByUsername(getCurrentUser().getUsername()).get();
+//		UserPrinciple user =  (UserPrinciple) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
 		Organization organizationSaved = organisationMapper
-				.OrganisationRequestDtoToOrganization(organizationRequestDto).setAdminId(userRepository.findByUsername(user.getUsername()).get().getUserId());
+				.OrganisationRequestDtoToOrganization(organizationRequestDto).setAdminId(userRepository.findByUsername(user.getUsername())
+						.get().getUserId()).setAppUser(user);
 		
 		return organisationMapper
 				.OrganisationToOrganizationResponseDto
@@ -82,12 +116,17 @@ public class OrganizationServiceImpl implements OrganizationService{
 
 	@Override
 	public OrganizationResponsDto update(OrganizationRequestDto organizationRequestDto, String orgaName) {
+		this.initCurrentUser();
+		if(this.getCheckAccessAdmin().checkAccess(getCurrentUser().getAdminId(), getCurrentUser())) {
 		
+			
+		}
 		Organization organizationUpdated = organizationRepository.findByName(orgaName);
 		organizationUpdated.setName(organizationRequestDto.getName())
 		.setAddress(organizationRequestDto.getAddress())
 		.setLogo(organizationRequestDto.getLogo())
 		.setNbSalaris(organizationRequestDto.getNbSalaris());
+		
 		
 		return organisationMapper
 				.OrganisationToOrganizationResponseDto
@@ -96,10 +135,18 @@ public class OrganizationServiceImpl implements OrganizationService{
 
 	@Override
 	public String remove(String orgaId) {
+		this.initCurrentUser();
+		
 		if (organizationRepository.existsById(orgaId)) {
-			organizationRepository.deleteById(orgaId);
-			
-			return "remove organization successfully";
+			Organization organization = organizationRepository.findById(orgaId).get();
+			if(checkAccessAdmin.checkAccess(organization.getAdminId(), this.currentUser)) {
+				organizationRepository.deleteById(orgaId);
+				return "remove organization successfully";
+			}
+			else {
+				return "The admin cannot remove this  organization!";
+			}
+
 		}else {
 			return "remove organization fail";
 		}
